remove redundant statistic collection intermediate classes
