+++ urbanstats/statistics/collections/segregation.py
@@ -1,5 +1,5 @@
 from urbanstats.geometry.segregation import compute_homogenity_statistics
-from urbanstats.statistics.statistic_collection import CensusStatisticsColection
+from urbanstats.statistics.statistic_collection import USAStatistics
 
 homogeneity_explanation = (
     "!TOOLTIP We define racial homogeneity as the average probability a person selecing "
@@ -7,7 +7,7 @@ homogeneity_explanation = (
 )
 
 
-class SegregationStatistics(CensusStatisticsColection):
+class SegregationStatistics(USAStatistics):
     version = 3
 
     base_stats = {
